+++ RPG_autoBattler/Effects/NinjaBleedEffect.cs
@@ -10,7 +10,7 @@ namespace RPG_autoBattler
 
         public void Trigger(TriggerType triggerType, Character attacker, Character victim, float[] specValue)
         {
-            if ((triggerType == TriggerType.EndTurn) && (specValue[0] == 0) && (TurnsLeft > 0))
+            if ((triggerType == TriggerType.EndTurn) && ((int)specValue[0] == 0) && (TurnsLeft > 0))
             {
                 Console.WriteLine($"{attacker} is bleeding! {TurnsLeft} turns left!");
                 attacker.CurHP -= DamagePercent * attacker.MaxHP;
